+++ stormpulse/logging/shipper.py
@@ -9,7 +9,12 @@ from typing import Any
 
 from stormpulse.config import LogGroupConfig
 from stormpulse.logging.parsers import PARSERS
-from stormpulse.logging.tailer import DockerTailer, LogTailer, StreamingDockerTailer
+from stormpulse.logging.tailer import (
+    DockerTailer,
+    JournaldTailer,
+    LogTailer,
+    StreamingDockerTailer,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -40,7 +45,7 @@ class LogShipper:
     def __init__(
         self,
         group: LogGroupConfig,
-        tailer: LogTailer | DockerTailer | StreamingDockerTailer,
+        tailer: LogTailer | DockerTailer | StreamingDockerTailer | JournaldTailer,
     ) -> None:
         self._group = group
         self._tailer = tailer
@@ -58,7 +63,7 @@ class LogShipper:
         return self._group.parser
 
     @property
-    def tailer(self) -> LogTailer | DockerTailer | StreamingDockerTailer:
+    def tailer(self) -> LogTailer | DockerTailer | StreamingDockerTailer | JournaldTailer:
         return self._tailer
 
     @property